refactor(P2.5 sub-slice A): move CTE-expression rewriters into render_plan/cte_rewrite.rs
Phase-2 §5.1 module move, first sub-slice of the (large) cte_rewrite group.
Extracted the self-contained CTE-expression-rewriting cluster verbatim from
plan_builder_utils.rs to a new render_plan/cte_rewrite.rs:
rewrite_operator_application_for_cte / rewrite_operator_application_for_cte_join /
rewrite_render_expr_for_cte_simple / rewrite_render_expr_for_cte_operand
(old lines 342-453). `pub(crate) use` re-exports for the externally-called
..._for_cte (join_builder) and the internally-called ..._for_cte_join (2 sites).

The only non-verbatim change: widened rewrite_operator_application_for_cte_join
from `fn` to `pub(crate) fn` so the re-export reaches it (documented at the
site) — same transition-visibility pattern as P2.1/P2.2.

P2.5 is sub-sliced because its §5.1 home is ~13 fns across 3 scattered bands
(much larger/entangled than P2.1-P2.4, and some listed fns belong to P2.6's
giant builders). Remaining P2.5: remap_cte_names_*, rewrite_logical_expr_cte_refs,
update_graph_joins_cte_refs, the P2.4-deferred alias walkers, D2 dedup.

Gates: 1612 lib + 529 integration (incl. corpus_sweep + sql_golden) + 1 ratchet
+ 7 unit, 0 failures; byte-identical (moved-fn bodies diff-clean vs main modulo
the one visibility line, no golden churn, ratchet net-zero); fmt/clippy clean;
warning-free --tests build.
