+++ docs/design/PRIORITIES.md
@@ -143,7 +143,7 @@ P-3). Latent finding filed in-report: `has_with_clause_in_graph_rel` is
 duplicated (utils + helpers) with a DIFFERENT semantic — a future consolidation
 candidate, not touched here (§8.3 no-drive-by).
 
-### P-3 — Phase 2 module moves (P2.1 → P2.6, in order)  ◐ (P2.1–P2.4 merged — P2.5 next)
+### P-3 — Phase 2 module moves (P2.1 → P2.6, in order)  ◐ (P2.1–P2.4 merged; P2.5 in progress, sub-sliced)
 The dead-code sweep shrank plan_builder_utils.rs to ~14.5K lines. §5.1 moves are
 now underway. Pure groups first (vlp_rewrite →
 pattern_comprehension_sql → clause_extractors → plan_predicates →
@@ -169,7 +169,15 @@ byte-identical goldens + corpus, ratchet net-zero. Scoped down from §5.1's
 premise: the private fresh-scan/with-exported alias walkers named there are
 coupled to P2.5's cte_rewrite fns, so they ride with P2.5 (§8.3 no-drive-by); the
 `plan_builder_helpers` `has_with_clause_in_graph_rel` D-cluster copy stays
-untouched. **Next: P2.5 cte_rewrite move.**
+untouched. **P2.5 (cte_rewrite move) in progress — sub-slice A delivered**: the
+CTE-expression-rewriting cluster (`rewrite_operator_application_for_cte`/`_join`
++ `rewrite_render_expr_for_cte_simple`/`_operand`) extracted to
+`render_plan/cte_rewrite.rs`, `pub(crate)` re-exports, byte-identical (one
+`fn`→`pub(crate) fn` visibility widening for the re-export), ratchet net-zero.
+P2.5 is being sub-sliced because its §5.1 home is ~13 fns across 3 scattered
+bands (much larger/entangled than P2.1–P2.4); remaining: `remap_cte_names_*`,
+`rewrite_logical_expr_cte_refs`, `update_graph_joins_cte_refs`, the deferred P2.4
+alias walkers, D2 dedup. **Next: P2.5 sub-slice B.**
 
 ### P-4 — Phase 4 §7.2: forward resolution through CTE scope  ◐ (F0+F1+F1b/#602+#662+F2a+F3+F4 done; F2b/F5 and F1b residue open)
 **Concrete staged plan written: `docs/design/FORWARD_RESOLUTION_PLAN.md`.** It
@@ -253,6 +261,24 @@ after P-2 merges), 1× P-5 S1. Re-balance here, in writing, not ad hoc.
 
 ## 4. Merge log (newest first — append on merge)
 
+- 2026-07-27: **P2.5 sub-slice A — cte_rewrite (expression-rewriting cluster)**
+  (P-3 refactor lane) — the self-contained CTE-expression-rewriting cluster
+  (`rewrite_operator_application_for_cte`, `rewrite_operator_application_for_cte_join`,
+  `rewrite_render_expr_for_cte_simple`, `rewrite_render_expr_for_cte_operand`, old
+  lines 342–453) extracted to a new `render_plan/cte_rewrite.rs`. `pub(crate) use`
+  re-exports for the externally-called `..._for_cte` (join_builder) and the
+  internally-called `..._for_cte_join` (2 sites in plan_builder_utils). The only
+  non-verbatim change is widening `rewrite_operator_application_for_cte_join` from
+  `fn` to `pub(crate) fn` so the re-export can reach it (documented at the site) —
+  same transition-visibility pattern as P2.1/P2.2. **P2.5 is being sub-sliced**
+  because its §5.1 home is ~13 fns across 3 scattered bands (much larger/entangled
+  than P2.1–P2.4, and some listed fns actually belong to P2.6's giant builders);
+  remaining P2.5: `remap_cte_names_*`, `rewrite_logical_expr_cte_refs`,
+  `update_graph_joins_cte_refs`, the P2.4-deferred alias walkers, D2 dedup. Gates:
+  1612 lib + 529 integration (incl. `corpus_sweep` + sql_golden) + 1 ratchet + 7
+  unit, 0 failures; **byte-identical** (moved-fn bodies diff-clean vs main modulo
+  the one visibility line, no golden churn, ratchet net-zero); fmt/clippy clean;
+  warning-free `--tests` build. Next: P2.5 sub-slice B.
 - 2026-07-27: **P2.4 — plan_predicates module move** (P-3 refactor lane) — the
   WITH-detection predicate group (`has_with_clause_in_tree`,
   `has_with_clause_in_graph_rel`, `plan_contains_with_clause`, old lines
